+++ src/screens/AccountSettings/components/DropdownToggleButton.tsx
@@ -1,0 +1,13 @@
+import { DropdownToggleButtonProps } from '../../../utilities/types/AccountSettingsInterface'
+import { handleToggleDropdown } from '../helper/dropdownHelpers'
+import styles from '../css/AccountSettings.module.css'
+
+export const DropdownToggleButton = ({ active, setting, tempModes, dropdownModes, setDropdownModes }: DropdownToggleButtonProps) => {
+
+  return (
+    <div onClick={() => handleToggleDropdown({ setting, tempModes, dropdownModes, setDropdownModes })}>
+      <button className={active ? styles.active_arrow : styles.arrow}>&#9002;</button>
+    </div>
+  )
+
+}
