+++ src/screens/AccountSettings/css/AccountSettings.module.css
@@ -1,9 +1,3 @@
-* {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
-}
-
 .account_settings_container {
   align-items: flex-start;
   display: flex;
